Inline the ModelIO interface
