+++ src/lightning/pytorch/core/mixins/hparams_mixin.py
@@ -17,8 +17,10 @@ import types
 from argparse import Namespace
 from typing import Any, List, MutableMapping, Optional, Sequence, Union
 
-from lightning.pytorch.core.saving import ALLOWED_CONFIG_TYPES, PRIMITIVE_TYPES
 from lightning.pytorch.utilities.parsing import AttributeDict, save_hyperparameters
+
+_PRIMITIVE_TYPES = (bool, int, float, str)
+_ALLOWED_CONFIG_TYPES = (AttributeDict, MutableMapping, Namespace)
 
 
 class HyperparametersMixin:
@@ -123,9 +125,9 @@ class HyperparametersMixin:
             hp = vars(hp)
         if isinstance(hp, dict):
             hp = AttributeDict(hp)
-        elif isinstance(hp, PRIMITIVE_TYPES):
-            raise ValueError(f"Primitives {PRIMITIVE_TYPES} are not allowed.")
-        elif not isinstance(hp, ALLOWED_CONFIG_TYPES):
+        elif isinstance(hp, _PRIMITIVE_TYPES):
+            raise ValueError(f"Primitives {_PRIMITIVE_TYPES} are not allowed.")
+        elif not isinstance(hp, _ALLOWED_CONFIG_TYPES):
             raise ValueError(f"Unsupported config type of {type(hp)}.")
         return hp
